Refactore storage/kvstore directory
