--- storage/kvstore/filesystemstore/mbed_lib.json
@@ -1,3 +1,0 @@
-{
-    "name": "filesystemstore"
-}
