--- storage/kvstore/global_api/mbed_lib.json
@@ -1,3 +1,0 @@
-{
-    "name": "kv-global-api"
-}
