--- storage/kvstore/kv_map/mbed_lib.json
@@ -1,3 +1,0 @@
-{
-    "name": "kv-map"
-}
