--- storage/kvstore/tdbstore/mbed_lib.json
@@ -1,3 +1,0 @@
-{
-    "name": "tdbstore"
-}
